solve Subsection error

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection");
 
 exports.createSection = async (req, res) => {
     try {
@@ -54,14 +55,23 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId, courseId } = req.body;
 		const section = await Section.findByIdAndUpdate(
 			sectionId,
 			{ sectionName },
 			{ new: true }
 		);
+		const course = await Course.findById(courseId)
+		.populate({
+		  path: "courseContent",
+		  populate: {
+			path: "subSection",
+		  },
+		})
+		.exec()
 		res.status(200).json({
 			success: true,
+			data: course,
 			message: section,
 		});
 	} catch (error) {
@@ -69,6 +79,7 @@ exports.updateSection = async (req, res) => {
 		res.status(500).json({
 			success: false,
 			message: "Internal server error",
+			error: error.message,
 		});
 	}
 };
@@ -77,8 +88,34 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
     try {
 
-        const { sectionId } = req.params;
-		await Section.findByIdAndDelete(sectionId);
+        const { sectionId, courseId } = req.params;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+			  courseContent: sectionId,
+			},
+		  })
+		  const section = await Section.findById(sectionId)
+		  console.log(sectionId, courseId)
+		  if (!section) {
+			return res.status(404).json({
+			  success: false,
+			  message: "Section not found",
+			})
+		  }
+		  // Delete the associated subsections
+		  await SubSection.deleteMany({ _id: { $in: section.subSection } })
+	  
+		  await Section.findByIdAndDelete(sectionId)
+	  
+		  // find the updated course and return it
+		  const course = await Course.findById(courseId)
+			.populate({
+			  path: "courseContent",
+			  populate: {
+				path: "subSection",
+			  },
+			})
+			.exec()
 
 		res.status(200).json({
 			success: true,
